changes on  warmUpWorkers and keepAlive calculations
now the config args has priority when setting this flags, except when working with an array  of workers. When working with an array of workers they should be keep alive since there is no way of restoring them.

now you can warm up to the max amount of workers according to the configuration and keep those workers alive if neede