updating easy-cancelable-promise
updating easy-cancelable-promise

the pair library easy-cancelable-promise was updated, now the library allows modularity so we are able to export only what is need.

this fix the problem of not having abort controller on node environments

avoid compile licenses
adding fixes to support modular imports
Adding error handlers on onMessage and callbacks executions.
Error handling was added in order to prevent messages with errors to stay permanently on the workerMessages.

if a non controlled error occurs the message gets deleted and the error re-throws